move cudd wrapper to include and rename
